feat:subcategory ubah database dan tambah alert

+++ resources/views/pages/data-master/sub-category/_function/function.blade.php
@@ -1,7 +1,7 @@
 <script>
     // SELECT2
     $(document).ready(function() {
-        $('#cat_id').select2({
+        $('#categories_id').select2({
             dropdownParent: $('#subCategory-model'),
             placeholder: 'Pilih...'
         })
@@ -28,25 +28,45 @@
                 $('#modalHeader').html("Edit Sub Kategori");
                 $('#subCategory-model').modal('show');
                 $('#id').val(res.id);
+                $('#categories_id').val(res.categories_id);
                 $('#kode_sub_kategori').val(res.kode_sub_kategori);
                 $('#nama_sub_kategori').val(res.nama_sub_kategori);
-                $('#categories_id').val(res.categories_id);
             }
         });
     }
 
     function deleteFunc(id) {
         var id = id;
-        $.ajax({
-            type: "DELETE",
-            url: "{{ route('sub-category.destroy') }}",
-            data: {
-                id: id
-            },
-            dataType: 'json',
-            success: function(res) {
-                var oTable = $('#myTable').dataTable();
-                oTable.fnDraw(false);
+        Swal.fire({
+            title: 'Apakah Anda yakin?',
+            text: "Data yang dihapus tidak dapat dikembalikan!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: '#3085d6',
+            confirmButtonText: 'Ya, Hapus!',
+            cancelButtonText: 'Batal'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    type: "DELETE",
+                    url: "{{ route('sub-category.destroy') }}",
+                    data: {
+                        id: id
+                    },
+                    dataType: 'json',
+                    success: function(res) {
+                        var oTable = $('#myTable').dataTable();
+                        oTable.fnDraw(false);
+                        Swal.fire({
+                            icon: 'success',
+                            title: 'Success',
+                            text: res.message,
+                            timer: 2000,
+                            showConfirmButton: false
+                        })
+                    }
+                });
             }
         });
     }
@@ -62,15 +82,35 @@
             contentType: false,
             processData: false,
             success: (response) => {
+                // console.log(response);
                 $("#subCategory-model").modal('hide');
                 var oTable = $('#myTable').dataTable();
                 oTable.fnDraw(false);
                 $("#btn-save").html('Submit');
                 $("#btn-save").attr("disabled", false);
+                if (response.success == true) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Success',
+                        text: response.message,
+                        timer: 1000,
+                        showConfirmButton: false
+                    });
+                } else if (response.success == false) {
+                    Swal.fire({
+                        icon: 'error',
+                        title: 'Failed',
+                        text: response.message,
+                        timer: 2000,
+                        showConfirmButton: false
+                    });
+                }
             },
-            error: function(data) {
-                console.log(data);
+            error: function(xhr, status, error) {
+                var errorMessage = xhr.status + ': ' + xhr.statusText;
+                console.error('Error - ' + errorMessage);
             }
+
         });
     });
     // CRUD
